refactor(extension): background.test.ts を機能ドメインごとに分割・整理 (#487)
